chore(suite-sync): rename file
# Conflicts:
#	suite-common/suite-sync/src/turnOnSuiteSync.ts
